+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - [#57](https://github.com/picimako/mockitools/issues/57): Added an inspection to report arguments passed into `Mockito.mock()` and `Mockito.spy()` which are designed to determine the mock type based on
 the variable's type they are assigned to, and not by the type passed into them. See [Mocking/spying without specifying class](https://javadoc.io/doc/org.mockito/mockito-core/latest/org/mockito/Mockito.html#mock_without_class).
 - [#58](https://github.com/picimako/mockitools/issues/58): Added an inspection to report `MockSettings`-based mock creations that have convenience methods or simpler variants to create the mock.
+- [#59](https://github.com/picimako/mockitools/issues/59): Added an intention action to expand simple mock and spy creation calls to use specific `MockSettings`,
+for example `Mockito.mock(<type>, <answer>)` is converted to `Mockito.mock(<type>, Mockito.withSettings().defaultAnswer(<answer>))`.   
 
 ### Changed 
 - Plugin configuration updates, and code improvements specific to migrating from Java 11 to 17.